qos: Remove pm_qos_update_request_timeout() API.
Using a timeout for a PM QoS request can lead to disastrous results on
power consumption. It's always possible to find a fixed scope in which a
PM QoS request should be applied, so timeouts aren't ever strictly
needed; they're usually just a lazy way of using PM QoS. Remove the API
so that it cannot be abused any longer.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Vishalcj17 <vishalcj@aospa.co>
Change-Id: Ic9db66419d3a6758c0b0f18c7967619da6a975a3